Add a class for managing the music database.
This will allow us to manage the music database in one place. The idea
is that in the future, we can create a class for scanning files in the
music directory and have it send signals to this class to add and remove
songs. Also, it should allows us to connect the signals back to the QML
code to update the gui correctly.